Moved vector stuff out to math.h
