+++ Dark Samurai Warrior Reloaded/Dark Samurai Warrior Reloaded.vcxproj
@@ -133,6 +133,9 @@
   <ItemGroup>
     <Image Include="..\assets\guy.bmp" />
   </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="math.h" />
+  </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
   </ImportGroup>
