+++ Dark Samurai Warrior Reloaded/Dark Samurai Warrior Reloaded.vcxproj.filters
@@ -23,4 +23,9 @@
       <Filter>Resource Files</Filter>
     </Image>
   </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="math.h">
+      <Filter>Source Files</Filter>
+    </ClInclude>
+  </ItemGroup>
 </Project>
